Fix program flow to avoid silly bugs
It is always needed to read the next command unless there's an exception. Took me 10 minutes to figure out that in another command that was added in the chain of elif clauses, it was sent over and over again as a reply due to the next command not being read correctly.